fixed circleci building

+++ src/engine/shared/map.cpp
@@ -60,7 +60,7 @@ public:
 
 						if((TilemapCount / pTilemap->m_Width != pTilemap->m_Height) || (TilemapSize / (int)sizeof(CTile) != TilemapCount))
 						{
-							dbg_msg("engine", "map layer too big (%d * %d * %u causes an integer overflow)", pTilemap->m_Width, pTilemap->m_Height, sizeof(CTile));
+							dbg_msg("engine", "map layer too big (%d * %d * %u causes an integer overflow)", pTilemap->m_Width, pTilemap->m_Height, unsigned(sizeof(CTile)));
 							return false;
 						}
 						CTile *pTiles = static_cast<CTile *>(mem_alloc(TilemapSize, 1));
